% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure.R
\name{Figure}
\alias{Figure}
\alias{fig}
\alias{fig_id}
\title{Create a Figure caption}
\usage{
Figure(text = NULL, image = NULL, label = NULL, hold = NULL)

fig(..., hold = FALSE)

fig_id(label = NULL)
}
\arguments{
\item{text}{The text of the caption. It can contain Markdown formatting.}

\item{image}{The Markdown code to include a figure like
\verb{![alt](path/to/figure.png)}, see example.}

\item{label}{The label to use (not required if the function is run from
within an R chunk because the chunk label is used by default).}

\item{hold}{Should we save the caption for later use, and return invisibly?
Default value is \code{FALSE} if \verb{image=} is provided or the function is \emph{not}
called from within an R chunk.}

\item{...}{Same arguments as for \code{\link[=Figure]{Figure()}}.}
}
\value{
The caption with a code to number the figure is returned (invisibly, if
\code{hold = TRUE}).
}
\description{
A caption for a figure can be created in two ways: either directly when you
provide both \verb{text=} and \verb{image=}, or indirectly, when you call \code{\link[=Figure]{Figure()}}
within an R chunk with \verb{image=} not provided. In that case, the caption is
saved for later use, and \code{fig()} retrieves it when needed. The label of the
figure is automatically set to the label of the chunk.
}
\examples{
Figure("The official R logo.",
"
![R logo](https://www.r-project.org/logo/Rlogo.png){width='80\%'}
")
}
\author{
Philippe Grosjean
}
\concept{automatic numbering of items in documents}
\keyword{utilities}
