% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.io-package.R
\docType{package}
\name{form.io-package}
\alias{form.io}
\alias{form.io-package}
\title{form.io: 'SciViews' - Objects formatting for R Markdown outputs}
\description{
Supplements to 'rmarkdown' and 'knitr' to output nicely formatted
  tables, figures and equations in R Markdown documents with caption.
}
\details{
The SciViews 'form.io' package provides a series of functions to enhance the
way R objects are output in R Markdown documents. It focuses on scientific
documents, and as such, emphasizes with the creation of figures and tables
that are numbered and captioned. The 'bookdown' package provides such a
feature through special formats (e.g., \code{bookdown::html_document2} instead of
\code{html_document}), but the functions provided here provide the same
enhancement for \emph{all} R Markdown files. The \code{\link[=form]{form()}} generic function is
there to provide a decent output format for many R objects that works in
HTML, PDF, Word and Markdown, as well as, at the R console. Under the hood,
it reuses formatting functions from other packages like \code{\link[knitr:kable]{knitr::kable()}}, or
\code{\link[pander:pander]{pander::pander()}}, but it makes sure that outputs are well-formatted in
(almost) all cases when working with R Markdown documents.
}
\section{Important functions}{

\itemize{
\item \code{\link[=form]{form()}} to format a series of R objects for R Markdown output.
\item \code{\link[=Figure]{Figure()}} and \code{\link[=Table]{Table()}} to construct captions for numbered figures and
tables.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/form.io}
  \item \url{https://www.sciviews.org/form.io/}
  \item Report bugs at \url{https://github.com/SciViews/form.io/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
