% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\alias{form.default}
\alias{form.anova}
\alias{form.data.frame}
\alias{form.summary.lm}
\alias{form.lm}
\alias{form.summary.glm}
\alias{form.summary.nls}
\alias{form.nls}
\title{Form tabular and textual output of various R objects using markdown}
\usage{
form(x, ...)

\method{form}{default}(
  x,
  caption = attr(x, "caption"),
  digits = 2,
  ...,
  data = NULL,
  options = list(digits = digits, missing = "", table.style = "rmarkdown"),
  table_split = 120L,
  table_row_align = "left",
  lang = getOption("data.io_lang", "en")
)

\method{form}{anova}(
  x,
  caption = attr(x, "caption"),
  data = NULL,
  signif_stars = TRUE,
  ...,
  lang = getOption("data.io_lang", "en")
)

\method{form}{data.frame}(
  x,
  caption = attr(x, "caption"),
  data = NULL,
  labels = TRUE,
  units = TRUE,
  ...,
  lang = getOption("data.io_lang", "en")
)

\method{form}{summary.lm}(
  x,
  caption = attr(x, "caption"),
  data = NULL,
  summary = TRUE,
  digits = 3,
  signif_stars = TRUE,
  ...,
  lang = getOption("data.io_lang", "en")
)

\method{form}{lm}(
  x,
  caption = attr(x, "caption"),
  data = NULL,
  digits = 2,
  signif_stars = TRUE,
  ...,
  lang = getOption("data.io_lang", "en")
)

\method{form}{summary.glm}(
  x,
  caption = attr(x, "caption"),
  data = NULL,
  summary = TRUE,
  digits = 3,
  signif_stars = TRUE,
  ...,
  lang = getOption("data.io_lang", "en")
)

\method{form}{summary.nls}(
  x,
  caption = attr(x, "caption"),
  data = NULL,
  summary = TRUE,
  digits = 3,
  signif_stars = TRUE,
  ...,
  lang = getOption("data.io_lang", "en")
)

\method{form}{nls}(
  x,
  caption = attr(x, "caption"),
  data = NULL,
  params_labels,
  digits = 2,
  signif_stars = TRUE,
  ...,
  lang = getOption("data.io_lang", "en")
)
}
\arguments{
\item{x}{An R object}

\item{...}{Further arguments (depends on the class of \code{x})}

\item{caption}{Legend for tables (if \code{""}, none)}

\item{digits}{Number of digits to print}

\item{data}{A data frame or tibble to be used to extract labels and/units}

\item{options}{Pander options (see \code{\link[=panderOptions]{panderOptions()}})}

\item{table_split}{Number of characters to accept for the width of the table
before splitting it (\code{\link[=pander]{pander()}} functions)}

\item{table_row_align}{How to align table row names (\code{"left"} by default)}

\item{lang}{The language to use for the output. By default, it is the
\code{data.io_lang} options (same one as for the \code{data.io} package)}

\item{signif_stars}{For \code{anova} objects, do we write significance stars?}

\item{labels}{Should we use labels for column names?}

\item{units}{Should we add units to labels in column names?}

\item{summary}{Should we add a summary of the object?}

\item{params_labels}{The label to use for the parameters.}
}
\value{
A character string with the formatted results
}
\description{
Use rich-text formatting (R markdown) to output the content of
\R objects by using \code{\link[=kable]{kable()}} or \code{\link[=pander]{pander()}} functions. Output are
readable at the R console, inside inline elements in an R Notebook and in
HTML, LaTeX or Word documents. Moreover, they are translatable in different
languages \verb{lang =} argument, and labels and/or units for variables
can also be used in place of the name of these variables.
}
\examples{
library(data.io) # Even better to use SciViews::R()
options(data.io_lang = "FR")
trees <- read("trees", package = "datasets")
lm_1 <- lm(data = trees, volume ~ .)
form(trees)
form(head(trees), data = trees) # Try without data =
form(anova(lm_1), data = trees)
}
\seealso{
\code{\link[knitr:kable]{knitr::kable()}}, \code{\link[pander:pander]{pander::pander()}}
}
\author{
Philippe Grosjean (phgrosjean@sciviews.org)
}
\concept{rich-formatted outputs}
\keyword{utilities}
