% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_captioner.R
\name{install_captioner}
\alias{install_captioner}
\title{Install hook and default chunk values for numbered figures, tables and equations}
\usage{
install_captioner()
}
\value{
Nothing, the function is intended for its side-effect of installing the
required knitr hook for processing numbered and captioned items. It also sets
default values for two chunk options: \code{fig.cap = form.io::fig()} and
\code{out.extra = form.io::fig_id()}. Both these chunk options allow to
automatically build a caption from the last call to \code{\link[=Figure]{Figure()}} within the
R chunk where the plot is created.
}
\description{
\code{\link[=install_captioner]{install_captioner()}} is called when 'form.io' loads. It is not necessary
to call it directly, except if you want to refresh these data. The automatic
installation when the 'form.io' package load can be disabled if you use
\code{options(form.io_captioner = FALSE)} before loading the package. In that case
you can still install it later with \code{\link[=install_captioner]{install_captioner()}}.
}
\author{
Philippe Grosjean
}
\concept{automatic numbering of items in documents}
\keyword{utilities}
