% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{Table}
\alias{Table}
\alias{tab}
\alias{print.knitr_asis}
\title{Create a Table caption}
\usage{
Table(text = NULL, table = NULL, label = NULL, hold = NULL)

tab(..., hold = FALSE)

\method{print}{knitr_asis}(x, ..., sep = "\\n")
}
\arguments{
\item{text}{The text of the caption. It can contain Markdown formatting.}

\item{table}{The textual description of the table (in plain Markdown).}

\item{label}{The label to use (not required if the function is run from
within an R chunk because the chunk label is used by default).}

\item{hold}{Should we save the caption for later use, and return invisibly?
Default value is \code{FALSE} if \verb{table=} is provided or the function is \emph{not}
called from within an R chunk.}

\item{...}{Same arguments as for \code{\link[=Table]{Table()}}, or arguments to respective
\code{\link[=print]{print()}} methods.}

\item{x}{Any R object to print.}

\item{sep}{Separator to use for printing multiple strings.}
}
\value{
The caption with a code to number the table is returned (invisibly, if
\code{hold = TRUE}).
}
\description{
A caption for a table can be created in two ways: either directly when you
provide both \verb{text=} and \verb{table=}, or indirectly, when you call \code{\link[=Table]{Table()}}
within an R chunk with \verb{table=} not provided. In that case, the caption is
saved for later use, and \code{tab()} retrieves it when needed. The label of the
table is automatically set to the label of the chunk.
}
\examples{
Table("This is my caption", "
|  x  |  y  |
|:---:|:---:|
|  1  |  2  |
|  3  |  5  |
")
}
\author{
Philippe Grosjean
}
\concept{automatic numbering of items in documents}
\keyword{utilities}
